\name{randAssign}
\alias{randAssign}
\title{Random assignment of group sizes}
\usage{
randAssign(totalPopulation, groupSize1, ...)
}
\description{
Randomly assigns specified group sizes to a given population. Accepts 2 or more integers, first the total population and then 1 or more group sizes.
}
\examples{
randElem(144, 4, 12, 20)
}
